Wal related debanjan

+++ actors/broker_manager/src/__init__.py
@@ -1,15 +1,17 @@
 from flask import Flask
+from flask.json.provider import DefaultJSONProvider
 from flask_sqlalchemy import SQLAlchemy
-import threading
 import requests
+from threading import Thread
 from time import sleep
+import os
 
 app = Flask(__name__)
-
 app.config[
     "SQLALCHEMY_DATABASE_URI"
-] = "postgresql://postgres:admin@localhost:5432/distributed_queue"
+] = "postgresql://postgres:admin@localhost:5432/distributed_queue_1"
 db = SQLAlchemy(app)
+
 
 from db_models import *
 from src.models import MasterQueue
@@ -21,31 +23,43 @@ from src import views
 
 def health_checker():
     while True:
-        try:
-            response = requests.get("http://127.0.0.1:5000/")
-            if response.status_code != 200:
-                print("Not Responding...")
-            else:
-                print("Alive")
-        except Exception as e:
-            print("Not Responding...")
-            # print(f"ERROR: {str(e)}")
-            pass
+        brokers = master_queue.get_brokers()
+        for broker in brokers:
+            try:
+                broker_url = broker.get_base_url()
+                response = requests.get(broker_url)
+                assert(response.status_code == 200)
+                print(f"{broker} Alive")
+                if not broker.is_alive():
+                    with app.app_context():
+                        broker.update_running_status(True)
+            except Exception as e:
+                print(f"{broker} Not Responding...")
+                if broker.is_alive():
+                    with app.app_context():
+                        broker.update_running_status(False)
 
-        sleep(20)
+        sleep(2)
 
 
 with app.app_context():
-    db.drop_all()
-    db.create_all()
-    master_queue.fetch_from_db()
+    if os.getenv("DEBUG") == "true":
+        db.drop_all()
+        db.create_all()
+    else:
+        db.create_all()
 
-    print("Starting health check threaad")
-    health_check_daemon = threading.Thread(
+    # fetch if it is a read_only manager
+    master_queue.do_init_sync()
+    
+    print("Done")
+
+    # finally reflect the changes done in database into in-memory data structures
+    master_queue.fetch_from_db() 
+    
+    print("Starting health check thread")
+    health_check_daemon = Thread(
         target=health_checker,
-        args=(),
         daemon=True,
     )
-
     health_check_daemon.start()
-
